# central-sso/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SSO Authentication API",
        "description": "Single Sign-On API for multi-tenant authentication",
        "contact": {
            "name": "SSO Admin",
            "email": "admin@sso.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/users/{userId}/roles": {
            "get": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Get all roles for a user",
                "operationId": "80db853ae3da5c6ec92fbe7f0cfe5594",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tenant_id",
                        "in": "query",
                        "description": "Filter by tenant ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "include_global",
                        "in": "query",
                        "description": "Include global roles",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "role": {
                                                        "$ref": "#/components/schemas/RoleResponseDTO"
                                                    },
                                                    "tenant_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Assign a role to a user",
                "operationId": "e684531bf9fd2149269bc2ba8632055e",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignRoleRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role assigned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Role assigned successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Role already assigned"
                    },
                    "404": {
                        "description": "User or role not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Remove a role from a user",
                "operationId": "194107adc837850909fba4c12b238100",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssignRoleRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Role removed successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Role not assigned to user"
                    },
                    "404": {
                        "description": "User or role not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{userId}/roles/sync": {
            "put": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Sync roles for a user in a specific tenant",
                "operationId": "b8a420e8bdf861fa6700a69f48e45855",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "roles": {
                                        "description": "Array of role slugs to assign",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "admin",
                                            "manager"
                                        ]
                                    },
                                    "tenant_id": {
                                        "description": "Tenant ID (null for global roles)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Roles synced successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Roles synced successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{userId}/permissions": {
            "get": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Get all permissions for a user through roles",
                "operationId": "95ac0183a7931a73420841eccadcc284",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tenant_id",
                        "in": "query",
                        "description": "Filter by tenant ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PermissionResponseDTO"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "AssignRoleRequestDTO": {
                "required": [
                    "role_slug"
                ],
                "properties": {
                    "role_slug": {
                        "description": "Role slug to assign",
                        "type": "string",
                        "example": "manager"
                    },
                    "tenant_id": {
                        "description": "Tenant ID for tenant-specific role assignment (optional for global roles)",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "PermissionResponseDTO": {
                "properties": {
                    "id": {
                        "description": "Permission ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Permission name",
                        "type": "string",
                        "example": "View Users"
                    },
                    "slug": {
                        "description": "Permission slug",
                        "type": "string",
                        "example": "users.view"
                    },
                    "description": {
                        "description": "Permission description",
                        "type": "string",
                        "example": "Allows viewing user list and details"
                    },
                    "category": {
                        "description": "Permission category",
                        "type": "string",
                        "example": "users"
                    },
                    "is_system": {
                        "description": "Whether the permission is a system permission",
                        "type": "boolean",
                        "example": false
                    },
                    "roles_count": {
                        "description": "Number of roles with this permission",
                        "type": "integer",
                        "example": 3
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    }
                },
                "type": "object"
            },
            "RoleResponseDTO": {
                "properties": {
                    "id": {
                        "description": "Role ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Role name",
                        "type": "string",
                        "example": "Manager"
                    },
                    "slug": {
                        "description": "Role slug",
                        "type": "string",
                        "example": "manager"
                    },
                    "description": {
                        "description": "Role description",
                        "type": "string",
                        "example": "Can manage users and view reports"
                    },
                    "is_system": {
                        "description": "Whether the role is a system role",
                        "type": "boolean",
                        "example": false
                    },
                    "permissions": {
                        "description": "List of permissions assigned to this role",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionResponseDTO"
                        }
                    },
                    "users_count": {
                        "description": "Number of users with this role",
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-16T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}